move the v1 mapStatus into log to hide it away
